\name{MeshSimData}
\alias{MeshSimData}
\docType{data}
\title{MeSH Dataset}
\description{
    These contents data of the whole MeSH tree, as well as information contents for every node and term.

    This dataset will be auto loaded by the first invoked function of this package, if no other dataset is specified.  
}
